nmt_training/requirements.txt
torch>=1.7.0
OpenNMT-py>=2.0.0
sentencepiece>=0.1.96
sacrebleu>=2.0.0
pyyaml>=5.4.0
psutil>=5.8.0
GPUtil>=1.4.0
matplotlib>=3.3.0
numpy>=1.19.0
tqdm>=4.50.0
spacy>=3.4.0
nltk>=3.6.0
pandas>=1.3.0
scikit-learn>=1.0.0
